knowledge/requirements.txt
aniso8601==8.0.0
click==7.1.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-igraph==0.8.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
texttable==1.6.2
Werkzeug==1.0.1
